pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thordata-langchain-tools"
version = "0.2.0"
description = "LangChain tools powered by Thordata - SERP API, Web Scraping, and Proxy Network."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Thordata Team", email = "support@thordata.com"}
]
keywords = [
    "langchain",
    "thordata",
    "web scraping",
    "serp",
    "proxy",
    "ai",
    "llm",
    "agent",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "thordata-sdk>=0.4.0",
    "langchain-core>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
openai = [
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
"Homepage" = "https://github.com/Thordata/thordata-langchain-tools"
"Documentation" = "https://github.com/Thordata/thordata-langchain-tools#readme"
"Source" = "https://github.com/Thordata/thordata-langchain-tools"
"Tracker" = "https://github.com/Thordata/thordata-langchain-tools/issues"

[tool.setuptools.packages.find]
include = ["thordata_langchain_tools*"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F"]
ignore = ["E501"]